pyproject.toml
[tool.poetry]
name = "BotNum2"
version = "0.1.0"
description = ""
authors = ["realboy"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
newspaper3k = "^0.2.8"
sklearn = "^0.0"
discord = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
